+++ clogging/src/features/Admin/types.ts
@@ -1,0 +1,12 @@
+import { ReactNode } from 'react';
+
+export interface AdminLayoutProps {
+  children: ReactNode;
+}
+
+export interface MenuItem {
+  name: string;
+  icon: string;
+  activeIcon: string;
+  href: string;
+}
