+++ clogging/src/features/Admin/ui/MainContent.tsx
@@ -1,0 +1,29 @@
+'use client';
+import { ReactNode } from 'react';
+import { useTheme } from '@/shared/providers/theme';
+import { Card } from '@/shared/ui/common/Card';
+import { cn } from '@/shared/lib/utils';
+import { usePathname } from 'next/navigation';
+import { menuItems } from '../constants';
+
+interface MainContentProps {
+  children: ReactNode;
+}
+
+export const MainContent = ({ children }: MainContentProps) => {
+  const { isDarkMode } = useTheme();
+  const pathname = usePathname();
+  const currentMenuItem =
+    menuItems.find((item) => item.href === pathname) || menuItems[0];
+
+  return (
+    <div className="w-full lg:mt-4 flex-grow">
+      <Card
+        className={cn(`h-full ${isDarkMode ? 'bg-gray-800' : 'bg-[#F4F7FE]'}`)}
+      >
+        <h1 className="text-2xl font-bold mb-6">{currentMenuItem.name}</h1>
+        {children}
+      </Card>
+    </div>
+  );
+};
